Support to save decoder contexts in the middle
In some network applications, the connection may be lost before the
decoder at the destination being able to finish. It is always desired to
save decoder's context into a file so that the decoding can resume later
when the connection is re-established. This commit adds the save/restore
features to all of the four decoders.