Use MT19937 for pseudo-random number generations
PRNG is needed in grouping subgeneration in RAND GNC codes, as well
as in generating coding coefficients in non-binary precodes (i.e., during
building the bipartite graph). Previously we used the same PRNG as
ANSI-C, but unfortunately that PRNG has poor performance in terms of
period, which in particular affects GF(2) precoding for a large number
of source packets.